+++ src/gmm_gibbs.c
@@ -11,10 +11,10 @@ struct gmm_sufficient_statistic {
     unsigned int *ns; 
 
     // sum of data points in each component
-    double *comp_sums;
+    float *comp_sums;
 
     // sum of the squares of data points in each component
-    double *comp_sqsums;
+    float *comp_sqsums;
 };
 
 struct gmm_gibbs_state {
@@ -24,7 +24,7 @@ struct gmm_gibbs_state {
     // number of mixture components
     size_t k; 
 
-    double *data;
+    float *data;
 
     struct gmm_prior prior;
     struct gmm_params *params;
@@ -32,7 +32,7 @@ struct gmm_gibbs_state {
 };
 
 struct gmm_gibbs_state *
-alloc_gmm_gibbs_state(size_t n, size_t k, double *data, struct gmm_prior prior,
+alloc_gmm_gibbs_state(size_t n, size_t k, float *data, struct gmm_prior prior,
                       struct gmm_params *params)
 {
     struct gmm_gibbs_state *s = abort_malloc(sizeof(struct gmm_gibbs_state));
@@ -44,8 +44,8 @@ alloc_gmm_gibbs_state(size_t n, size_t k, double *data, struct gmm_prior prior,
     s->ss = (struct gmm_sufficient_statistic*)
         abort_malloc(sizeof(struct gmm_sufficient_statistic));
     s->ss->ns = (unsigned *) abort_calloc(k, sizeof(unsigned int));
-    s->ss->comp_sums = (double *) abort_calloc(k, sizeof(double));
-    s->ss->comp_sqsums = (double *) abort_calloc(k, sizeof(double));
+    s->ss->comp_sums = (float *) abort_calloc(k, sizeof(float));
+    s->ss->comp_sqsums = (float *) abort_calloc(k, sizeof(float));
     return s;
 }
 
@@ -61,8 +61,8 @@ void free_gmm_gibbs_state(struct gmm_gibbs_state *state)
 void clear_sufficient_statistic(struct gmm_gibbs_state *state)
 {
     memset(state->ss->ns, 0, state->k * sizeof(unsigned int));
-    memset(state->ss->comp_sums, 0, state->k * sizeof(double));
-    memset(state->ss->comp_sqsums, 0, state->k * sizeof(double));
+    memset(state->ss->comp_sums, 0, state->k * sizeof(float));
+    memset(state->ss->comp_sqsums, 0, state->k * sizeof(float));
 }
 
 void update_sufficient_statistic(struct gmm_gibbs_state *state)
@@ -70,7 +70,7 @@ void update_sufficient_statistic(struct gmm_gibbs_state *state)
     // XXX XXX this is the function that needs to be accelerated.
     clear_sufficient_statistic(state);
     for(size_t i=0; i < state->n; i++) {
-        double x = state->data[i];
+        float x = state->data[i];
         unsigned int z = state->params->zs[i];
         state->ss->ns[z]++;
         state->ss->comp_sums[z] += x;
@@ -80,17 +80,17 @@ void update_sufficient_statistic(struct gmm_gibbs_state *state)
 
 void update_ws(struct gmm_gibbs_state *state)
 {
-    double dirichlet_param[state->k];
+    float dirichlet_param[state->k];
     vec_add_ud(dirichlet_param, state->ss->ns, state->params->weights, state->k);
     dirichlet(state->params->weights, dirichlet_param, state->k);
 }
 
 void update_means(struct gmm_gibbs_state *state)
 {
-    double k = 1/state->prior.means_var_prior,
+    float k = 1/state->prior.means_var_prior,
            zeta = state->prior.means_mean_prior, mean, var;
     for(int j=0; j < state->k; j++) {
-        double sum_xs = state->ss->comp_sums[j], ns = state->ss->ns[j],
+        float sum_xs = state->ss->comp_sums[j], ns = state->ss->ns[j],
                sigma2 = state->params->vars[j];
                mean = (k * zeta + sum_xs / sigma2) / (ns / sigma2 + k);
                var = 1/(ns / sigma2 + k);
@@ -100,10 +100,10 @@ void update_means(struct gmm_gibbs_state *state)
 
 void update_vars(struct gmm_gibbs_state *state)
 {
-    double alpha = state->prior.vars_shape_prior,
+    float alpha = state->prior.vars_shape_prior,
            beta = state->prior.vars_scale_prior, shape, scale;
     for(int j=0; j < state->k; j++) {
-        double sum_xs = state->ss->comp_sums[j],
+        float sum_xs = state->ss->comp_sums[j],
                sqsum_xs = state->ss->comp_sqsums[j],
                mu = state->params->means[j], ns = state->ss->ns[j];
                shape = alpha + ns/2;
@@ -114,9 +114,9 @@ void update_vars(struct gmm_gibbs_state *state)
 
 void update_zs(struct gmm_gibbs_state *state)
 {
-    double weights[state->k], mu, sigma2;
+    float weights[state->k], mu, sigma2;
     for(int i=0; i < state->n; i++) {
-        double x = state->data[i];
+        float x = state->data[i];
         for(int j=0; j < state->k; j++) {
             mu = state->params->means[j];
             sigma2 = state->params->vars[j];
